Move FullnodeController to its correct folder.
