+++ src/Stratis.Bitcoin.Features.RPC/Controllers/RPCController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.Extensions.Logging;
 using NBitcoin.RPC;
+using Stratis.Bitcoin.Controllers;
 using Stratis.Bitcoin.Utilities;
 using Stratis.Bitcoin.Utilities.JsonErrors;
